+++ internal/credentials/vault_other.go
@@ -4,6 +4,6 @@ package credentials
 
 import "github.com/zalando/go-keyring"
 
-func save(key string) error { return keyring.Set(Service, Account, key) }
-func read() (string, error) { return keyring.Get(Service, Account) }
-func remove() error         { return keyring.Delete(Service, Account) }
+func save(service, account, key string) error      { return keyring.Set(service, account, key) }
+func read(service, account string) (string, error) { return keyring.Get(service, account) }
+func remove(service, account string) error         { return keyring.Delete(service, account) }